This change removes the FileName attribute, providing a significant
performance boost. This cuts average run time in half again on top of the
first set of performance improvements.

Before:

```
Analysis Stats
================
Report Generated by      : ExPerfAnalyzer.ps1 v1.0
Written by               : Matthew Huynh (mahuynh@microsoft.com) & David Paulson (dpaul@microsoft.com
Generated On             : 7/22/2017 8:33 AM
Total Counters Processed : 517
Total Samples            : 744,997
Total Processing Time    : 61.9s
Samples Processed/sec    : 12,034.04878s
Proc. Time Per Sample    : 0.00008s

Script took 61.9250283 seconds to complete
[7/22/2017 8:33:19 AM] : Get-PerformanceDataFromFileLocal  took 19.98364 seconds to complete.
[7/22/2017 8:33:19 AM] : Add-CountersToAnalyzeToObject took 0.6200232 seconds to complete.
[7/22/2017 8:33:19 AM] : Analyze-DataOfObject took 24.805635 seconds to complete
[7/22/2017 8:33:19 AM] : Output-QuickSummaryDetails took 0.5941671 seconds to complete
[7/22/2017 8:33:19 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject took 4.449223 seconds to complete
[7/22/2017 8:33:19 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject Total Group path took 4.2082968 seconds to complete.
[7/22/2017 8:33:19 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject build server object took 0.2322907 seconds to complete.
[7/22/2017 8:33:19 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject grouping objects for server build took 0.085508 seconds to complete
[7/22/2017 8:33:19 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject build server time finder took 0.116268 seconds to complete.
```

After:

```
Analysis Stats
================
Report Generated by      : ExPerfAnalyzer.ps1 v1.0
Written by               : Matthew Huynh (mahuynh@microsoft.com) & David Paulson (dpaul@microsoft.com
Generated On             : 7/22/2017 8:38 AM
Total Counters Processed : 517
Total Samples            : 744,997
Total Processing Time    : 30.7s
Samples Processed/sec    : 24,275.35271s
Proc. Time Per Sample    : 0.00004s

Script took 30.7034536 seconds to complete
[7/22/2017 8:38:41 AM] : Get-PerformanceDataFromFileLocal  took 12.7906639 seconds to complete.
[7/22/2017 8:38:41 AM] : Add-CountersToAnalyzeToObject took 0.1347331 seconds to complete.
[7/22/2017 8:38:41 AM] : Analyze-DataOfObject took 7.2726312 seconds to complete
[7/22/2017 8:38:41 AM] : Output-QuickSummaryDetails took 0.2394321 seconds to complete
[7/22/2017 8:38:41 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject took 1.3492061 seconds to complete
[7/22/2017 8:38:41 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject Total Group path took 1.2713126 seconds to complete.
[7/22/2017 8:38:41 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject build server object took 0.0746006 seconds to complete.
[7/22/2017 8:38:41 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject grouping objects for server build took 0.0083491 seconds to complete
[7/22/2017 8:38:41 AM] : Convert-PerformanceCounterSampleObjectToServerPerformanceObject build server time finder took 0.043734 seconds to complete.
```